+++ .idea/misc.xml
@@ -33,6 +33,7 @@
         <entry key="..\:/Users/yechi/AndroidStudioProjects/MyApplication/chats/app/src/main/res/layout/card_chats.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/yechi/AndroidStudioProjects/MyApplication/chats/app/src/main/res/layout/chat_page.xml" value="0.24010416666666667" />
         <entry key="..\:/Users/yechi/AndroidStudioProjects/MyApplication/chats/app/src/main/res/layout/chats_recyclerview.xml" value="0.25416666666666665" />
+        <entry key="..\:/Users/yechi/AndroidStudioProjects/MyApplication/chats/app/src/main/res/layout/fragment.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/yechi/AndroidStudioProjects/MyApplication/chats/app/src/main/res/layout/toolbarchat.xml" value="0.13854166666666667" />
       </map>
     </option>
