+++ app/src/main/java/com/example/chats/client/Message.java
@@ -1,4 +1,4 @@
-package com.example.chats;
+package com.example.chats.client;
 
 public class Message {
