+++ app/src/main/java/com/example/chats/client/RecyclerViewAllChats.java
@@ -1,4 +1,4 @@
-package com.example.chats;
+package com.example.chats.client;
 
 import android.content.Context;
 import android.content.Intent;
@@ -8,12 +8,14 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Filter;
 import android.widget.ImageButton;
-import android.widget.ImageView;
 import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.cardview.widget.CardView;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.chats.R;
+
 import java.util.ArrayList;
 import java.util.List;
