+++ app/src/main/java/com/example/chats/client/RecyclerViewChat.java
@@ -1,4 +1,4 @@
-package com.example.chats;
+package com.example.chats.client;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -9,6 +9,9 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.chats.R;
+
 import java.util.ArrayList;
 import java.util.List;
