+++ app/src/main/java/com/example/chats/client/SettingsActivity.java
@@ -1,9 +1,11 @@
-package com.example.chats;
+package com.example.chats.client;
 
 import android.os.Bundle;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.preference.PreferenceFragmentCompat;
+
+import com.example.chats.R;
 
 public class SettingsActivity extends AppCompatActivity {
