+++ app/src/main/java/com/example/chats/firebase/Example.java
@@ -1,0 +1,4 @@
+package com.example.chats.firebase;
+
+public class Example {
+}
