+++ app/src/main/res/layout/fragment.xml
@@ -4,11 +4,13 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
+    android:gravity="center"
     android:padding="10sp">
 
     <LinearLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+
         android:gravity="center">
 
         <TextView
@@ -20,7 +22,7 @@
             <Button
                 android:id="@+id/picture"
                 android:layout_margin="10dp"
-                android:layout_width="wrap_content"
+                android:layout_width="60dp"
                 android:layout_height="wrap_content"
                 android:background="@drawable/ic_baseline_image_24"
                 android:textSize="15sp"/>
@@ -28,7 +30,7 @@
             <Button
                 android:id="@+id/camera"
                 android:layout_margin="10dp"
-                android:layout_width="wrap_content"
+                android:layout_width="60dp"
                 android:layout_height="wrap_content"
                 android:background="@drawable/ic_baseline_photo_camera_24"
                 android:textSize="15sp"/>
